+++ seacatauth/authn/handler.py
@@ -10,6 +10,7 @@ import aiohttp.web
 import urllib.parse
 import jwcrypto.jwk
 
+import seacatauth.exceptions
 from ..audit import AuditCode
 from ..cookie import set_cookie, delete_cookie
 from ..decorators import access_control
@@ -32,29 +33,31 @@ class AuthenticationHandler(object):
 	def __init__(self, app, authn_svc):
 		self.App = app
 		self.AuthenticationService = authn_svc
-		self.CredentialsService = app.get_service('seacatauth.CredentialsService')
-		self.SessionService = app.get_service('seacatauth.SessionService')
-		self.CookieService = app.get_service('seacatauth.CookieService')
-		self.AuditService = app.get_service('seacatauth.AuditService')
+		self.CredentialsService = app.get_service("seacatauth.CredentialsService")
+		self.SessionService = app.get_service("seacatauth.SessionService")
+		self.CookieService = app.get_service("seacatauth.CookieService")
+		self.AuditService = app.get_service("seacatauth.AuditService")
 		self.BatmanService = app.BatmanService
-		self.CommunicationService = app.get_service('seacatauth.CommunicationService')
+		self.CommunicationService = app.get_service("seacatauth.CommunicationService")
 
 		web_app = app.WebContainer.WebApp
-		web_app.router.add_put(r'/public/login.prologue', self.login_prologue)
-		web_app.router.add_put(r'/public/login/{lsid}', self.login)
-		web_app.router.add_put(r'/public/login/{lsid}/smslogin', self.smslogin)
-		web_app.router.add_put(r'/public/login/{lsid}/webauthn', self.webauthn_login)
-		web_app.router.add_put(r'/public/logout', self.logout)
-		web_app.router.add_post("/impersonate", self.impersonate)
+		web_app.router.add_put(r"/public/login.prologue", self.login_prologue)
+		web_app.router.add_put(r"/public/login/{lsid}", self.login)
+		web_app.router.add_put(r"/public/login/{lsid}/smslogin", self.smslogin)
+		web_app.router.add_put(r"/public/login/{lsid}/webauthn", self.webauthn_login)
+		web_app.router.add_put(r"/public/logout", self.logout)
+		web_app.router.add_put("/impersonate", self.impersonate)
+		web_app.router.add_post("/impersonate", self.impersonate_and_redirect)
 
 		# Public endpoints
 		web_app_public = app.PublicWebContainer.WebApp
-		web_app_public.router.add_put(r'/public/login.prologue', self.login_prologue)
-		web_app_public.router.add_put(r'/public/login/{lsid}', self.login)
-		web_app_public.router.add_put(r'/public/login/{lsid}/smslogin', self.smslogin)
-		web_app_public.router.add_put(r'/public/login/{lsid}/webauthn', self.webauthn_login)
-		web_app_public.router.add_put(r'/public/logout', self.logout)
-		web_app_public.router.add_post("/impersonate", self.impersonate)
+		web_app_public.router.add_put(r"/public/login.prologue", self.login_prologue)
+		web_app_public.router.add_put(r"/public/login/{lsid}", self.login)
+		web_app_public.router.add_put(r"/public/login/{lsid}/smslogin", self.smslogin)
+		web_app_public.router.add_put(r"/public/login/{lsid}/webauthn", self.webauthn_login)
+		web_app_public.router.add_put(r"/public/logout", self.logout)
+		web_app_public.router.add_put("/impersonate", self.impersonate)
+		web_app_public.router.add_post("/impersonate", self.impersonate_and_redirect)
 
 	@asab.web.rest.json_schema_handler({
 		"type": "object",
@@ -297,15 +300,20 @@ class AuthenticationHandler(object):
 
 		delete_cookie(self.App, response)
 
-		# If the current session is impersonated, try to restore the original session
+		# If the current session is impersonated and the original session has a
+		# root cookie, try to restore the original session
 		if session.Authentication.ImpersonatorSessionId is not None:
 			try:
 				impersonator_session = await self.SessionService.get(session.Authentication.ImpersonatorSessionId)
-				print(impersonator_session)
-				set_cookie(self.App, response, impersonator_session)
 			except KeyError:
 				L.log(asab.LOG_NOTICE, "Impersonator session not found.", struct_data={
 					"sid": session.Authentication.ImpersonatorSessionId})
+			else:
+				if impersonator_session.Cookie is None:
+					# Case when the impersonation was started by an M2M session, which has no cookie
+					pass
+				else:
+					set_cookie(self.App, response, impersonator_session)
 
 		if self.BatmanService is not None:
 			response.del_cookie(self.BatmanService.CookieName)
@@ -385,13 +393,54 @@ class AuthenticationHandler(object):
 			login_key = None
 		return login_key
 
+	@asab.web.rest.json_schema_handler({
+		"type": "object",
+		"required": ["credentials_id"],
+		"properties": {
+			"credentials_id": {
+				"type": "string",
+				"description": "Credentials ID of the impersonation target."},
+			"expiration": {
+				"oneOf": [{"type": "string"}, {"type": "number"}],
+				"description":
+					"Expiration of the impersonated session. The value can be either the number of seconds "
+					"or a time-unit string such as '4 h' or '3 d'."}},
+		"example": {
+			"credentials_id": "mongodb:default:abc123def456",
+			"expiration": "5m"}
+	})
+	@access_control("authz:impersonate")
+	async def impersonate(self, request, *, json_data):
+		"""
+		Open a root session impersonated as a different user.
+		Response contains a Set-Cookie header with the new root session cookie.
+
+		Requires `authz:impersonate`.
+		"""
+		from_info = [request.remote]
+		ff = request.headers.get("X-Forwarded-For")
+		if ff is not None:
+			from_info.extend(ff.split(", "))
+
+		target_cid = json_data["credentials_id"]
+		if request.Session.Session.ParentSessionId is None:
+			impersonator_root_session = request.Session
+		else:
+			impersonator_root_session = await self.SessionService.get(request.Session.Session.ParentSessionId)
+
+		session = await self._impersonate(impersonator_root_session, from_info, target_cid)
+		response = asab.web.rest.json_response(request, {"result": "OK"})
+		set_cookie(self.App, response, session, cookie_domain=self.CookieService.RootCookieDomain)
+		return response
+
 
 	@access_control("authz:impersonate")
-	async def impersonate(self, request, *, credentials_id):
+	async def impersonate_and_redirect(self, request):
 		"""
-		Open a root session as a different user. Responds contains a Set-Cookie header with the new root session cookie.
-		This effectively overwrites user's current root cookie. Reference to current root session is kept in the
-		impersonated session. On logout, the original root cookie is set again.
+		Open a root session impersonated as a different user. Response contains a Set-Cookie header with the new
+		root session cookie and redirection to the authorize endpoint. This effectively overwrites user's current
+		root cookie. Reference to current root session is kept in the impersonated session.
+		On logout, the original root cookie is set again.
 
 		Requires `authz:impersonate`.
 		---
@@ -427,7 +476,6 @@ class AuthenticationHandler(object):
 		oidc_service = self.App.get_service("seacatauth.OpenIdConnectService")
 		client_service = self.App.get_service("seacatauth.ClientService")
 
-		# TODO: Restrict impersonation based on agent X target resource intersection
 		from_info = [request.remote]
 		ff = request.headers.get("X-Forwarded-For")
 		if ff is not None:
@@ -435,47 +483,12 @@ class AuthenticationHandler(object):
 
 		request_data = await request.post()
 		target_cid = request_data["credentials_id"]
-
 		if request.Session.Session.Type == "root":
 			impersonator_root_session = request.Session
 		else:
 			impersonator_root_session = await self.SessionService.get(request.Session.Session.ParentSessionId)
 
-		try:
-			session = await self.AuthenticationService.create_impersonated_session(impersonator_root_session, target_cid)
-		except Exception as e:
-			await self.AuditService.append(
-				AuditCode.IMPERSONATION_FAILED,
-				{
-					"impersonator_cid": credentials_id,
-					"impersonator_sid": impersonator_root_session.SessionId,
-					"target_cid": target_cid,
-					"fi": from_info,
-				}
-			)
-			raise e
-		else:
-			L.log(
-				asab.LOG_NOTICE,
-				"Impersonation successful.",
-				struct_data={
-					"impersonator_cid": credentials_id,
-					"impersonator_sid": impersonator_root_session.SessionId,
-					"target_cid": target_cid,
-					"target_sid": str(session.Session.Id),
-					"from_info": from_info,
-				}
-			)
-			await self.AuditService.append(
-				AuditCode.IMPERSONATION_SUCCESSFUL,
-				{
-					"impersonator_cid": credentials_id,
-					"impersonator_sid": impersonator_root_session.SessionId,
-					"target_cid": target_cid,
-					"target_sid": session.Session.Id,
-					"fi": from_info,
-				}
-			)
+		session = await self._impersonate(impersonator_root_session, from_info, target_cid)
 
 		client_dict = await client_service.get(request_data["client_id"])
 		query = {
@@ -496,3 +509,58 @@ class AuthenticationHandler(object):
 		)
 		set_cookie(self.App, response, session, cookie_domain=self.CookieService.RootCookieDomain)
 		return response
+
+	async def _impersonate(self, impersonator_root_session, impersonator_from_info, target_cid):
+		"""
+		Create a new impersonated session and log the event.
+		"""
+		# TODO: Restrict impersonation based on agent X target resource intersection
+		impersonator_cid = impersonator_root_session.Credentials.Id
+		try:
+			session = await self.AuthenticationService.create_impersonated_session(
+				impersonator_root_session, target_cid)
+		except seacatauth.exceptions.AccessDeniedError as e:
+			await self.AuditService.append(
+				AuditCode.IMPERSONATION_FAILED,
+				{
+					"impersonator_cid": impersonator_cid,
+					"impersonator_sid": impersonator_root_session.SessionId,
+					"target_cid": target_cid,
+					"fi": impersonator_from_info,
+				}
+			)
+			raise aiohttp.web.HTTPForbidden() from e
+		except Exception as e:
+			await self.AuditService.append(
+				AuditCode.IMPERSONATION_FAILED,
+				{
+					"impersonator_cid": impersonator_cid,
+					"impersonator_sid": impersonator_root_session.SessionId,
+					"target_cid": target_cid,
+					"fi": impersonator_from_info,
+				}
+			)
+			raise e
+		else:
+			L.log(
+				asab.LOG_NOTICE,
+				"Impersonation successful.",
+				struct_data={
+					"impersonator_cid": impersonator_cid,
+					"impersonator_sid": impersonator_root_session.SessionId,
+					"target_cid": target_cid,
+					"target_sid": str(session.Session.Id),
+					"from_info": impersonator_from_info,
+				}
+			)
+			await self.AuditService.append(
+				AuditCode.IMPERSONATION_SUCCESSFUL,
+				{
+					"impersonator_cid": impersonator_cid,
+					"impersonator_sid": impersonator_root_session.SessionId,
+					"target_cid": target_cid,
+					"target_sid": session.Session.Id,
+					"fi": impersonator_from_info,
+				}
+			)
+		return session